Move graph state wrapper component into state directory
