+++ src/components/admin/Billing/index.tsx
@@ -3,8 +3,8 @@ import { authenticatedRoutes } from 'app/routes';
 import CardWrapper from 'components/admin/Billing/CardWrapper';
 import DataByMonthGraph from 'components/admin/Billing/graphs/DataByMonthGraph';
 import DataByTaskGraph from 'components/admin/Billing/graphs/DataByTaskGraph';
+import GraphStateWrapper from 'components/admin/Billing/graphs/states/Wrapper';
 import TasksByMonth from 'components/admin/Billing/graphs/TasksByMonthGraph';
-import GraphStateWrapper from 'components/admin/Billing/graphs/Wrapper';
 import PaymentMethods from 'components/admin/Billing/PaymentMethods';
 import PricingTierDetails from 'components/admin/Billing/PricingTierDetails';
 import TenantOptions from 'components/admin/Billing/TenantOptions';
